+++ frontend/src/lib/components/popups/PlayerPopups.svelte
@@ -1,0 +1,11 @@
+<script lang="ts">
+    import PastGamePopup from "$lib/components/popups/PastGamePopup.svelte";
+    import PastPlayerPopup from "$lib/components/popups/PastPlayerPopup.svelte";
+    import ProfilePicturePopup from "$lib/components/popups/ProfilePicturePopup.svelte";
+    import FriendsPopup from "./FriendsPopup.svelte";
+</script>
+
+<ProfilePicturePopup />
+<PastGamePopup />
+<PastPlayerPopup />
+<FriendsPopup />