+++ .github/release-notes.md
@@ -1,6 +1,6 @@
 * Attempt to fetch missing summaries when viewing player history
 * Added green highlight to players in your game when viewing history
 * Fixed players not being clickable in past game popup
-* Added display for known friends when clicking on player in game
+* Added popup for viewing known friends of a player
 * Added automatic tag for steam friends
 * Added option to not have tags apply highlight